+++ platform/RELEASES.md
@@ -9,3 +9,13 @@ Based on:
 - [typescript v0.1.0] platform
 ### Releases
 - [NPM v0.1.0] https://www.npmjs.com/package/@codat/platform/v/0.1.0 - platform
+
+## 2023-09-07 00:04:02
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Common.yaml
+- Speakeasy CLI 1.79.0 (2.101.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.2.0] platform
+### Releases
+- [NPM v0.2.0] https://www.npmjs.com/package/@codat/platform/v/0.2.0 - platform